+++ CHANGELOG.md
@@ -44,6 +44,7 @@
  - fixed ui not showing 3p calibrations
  - experiment profiles start now use the unit_api directly. This may mitigate the issue where huey workers stampeding on each other when try to start many jobs.
  - fix `pio calibrations run ... -y` not saving as active.
+ - fix manual dosing in the UI
 
 ### 25.1.21